requirements.simple.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0

# Real-time features
channels==4.0.0

# Background tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File handling & Storage
Pillow==10.1.0
django-storages==1.14.2

# API Documentation
drf-spectacular==0.26.5

# Payment Processing
stripe==7.8.0

# AI/ML Integration
openai>=1.10.0,<2.0.0